# openapi.yaml
openapi: 3.0.3

info:
  title: Alerta Segura API
  description: API REST para reportar huecos y luminarias dañadas al municipio.
  version: 1.0.0

servers:
  - url: http://localhost:3000

paths:

  /reports:
    get:
      summary: Obtener reportes con paginación
      description: Obtiene una lista de reportes utilizando los parámetros page y limit.
      parameters:
        - name: page
          in: query
          required: false
          description: Número de página.
          schema:
            type: integer
            minimum: 1
            default: 1
            example: 1

        - name: limit
          in: query
          required: false
          description: Cantidad de reportes por página.
          schema:
            type: integer
            minimum: 1
            default: 5
            example: 2

      responses:
        '200':
          description: Lista de reportes obtenida correctamente

        '400':
          description: Parámetros de paginación inválidos

    post:
      summary: Crear un nuevo reporte
      description: Crea un reporte y requiere autenticación mediante Bearer Token.

      security:
        - bearerAuth: []

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Report'

      responses:
        '201':
          description: Reporte creado correctamente

        '400':
          description: Datos de entrada inválidos

        '401':
          description: No autorizado. Se requiere un token válido.


  /reports/{id}:

    get:
      summary: Obtener un reporte por ID

      parameters:
        - name: id
          in: path
          required: true
          description: Identificador del reporte.
          schema:
            type: integer
            example: 1

      responses:
        '200':
          description: Reporte encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'

        '404':
          description: Reporte no encontrado


    put:
      summary: Actualizar un reporte

      parameters:
        - name: id
          in: path
          required: true
          description: Identificador del reporte.
          schema:
            type: integer
            example: 1

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Report'

      responses:
        '200':
          description: Reporte actualizado correctamente

        '404':
          description: Reporte no encontrado


    delete:
      summary: Eliminar un reporte

      parameters:
        - name: id
          in: path
          required: true
          description: Identificador del reporte.
          schema:
            type: integer
            example: 1

      responses:
        '204':
          description: Reporte eliminado correctamente

        '404':
          description: Reporte no encontrado


components:

  securitySchemes:

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token

  schemas:

    Report:
      type: object

      properties:

        id:
          type: integer
          example: 1

        description:
          type: string
          example: Hueco grande en la vía

        location:
          type: string
          example: Bogotá, Calle 80

        photo:
          type: string
          example: foto.jpg

        status:
          type: string
          example: pending